StockPricesMonitoring-master/Solution/src/main/java/com/example/model/MyStock.java
package com.example.model;

import java.sql.Timestamp;

public class MyStock {
    private String symbol;
    private double price;
    String timeStamp;
    String name;

    public MyStock() {

    }

    public String getTimeStamp() {
        return timeStamp;
    }

    public void setTimeStamp(String timeStamp) {
        this.timeStamp = timeStamp;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSymbol() {
        return symbol;
    }
    public void setSymbol(String symbol) {
        this.symbol = symbol;
    }
    public double getPrice() {
        return price;
    }
    public void setPrice(double price) {
        this.price = price;
    }

}
